apps/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
alembic
pydantic
email-validator
pydantic-settings
python-dotenv
stripe
httpx
pytest
pyjwt[crypto]
